+++ cmd/uptrace/command/ch.go
@@ -9,6 +9,7 @@ import (
 	"github.com/uptrace/go-clickhouse/ch/chschema"
 	"github.com/uptrace/go-clickhouse/chmigrate"
 	"github.com/uptrace/uptrace/pkg/bunapp"
+	"github.com/uptrace/uptrace/pkg/uptracebundle"
 	"github.com/urfave/cli/v2"
 	"go.uber.org/zap"
 )
@@ -22,7 +23,7 @@ func NewCHCommand(migrations *chmigrate.Migrations) *cli.Command {
 				Name:  "wait",
 				Usage: "wait until ClickHouse is up and running",
 				Action: func(c *cli.Context) error {
-					ctx, app, err := bunapp.StartCLI(c)
+					ctx, app, err := uptracebundle.StartCLI(c)
 					if err != nil {
 						return err
 					}
@@ -50,7 +51,7 @@ func NewCHCommand(migrations *chmigrate.Migrations) *cli.Command {
 				Name:  "init",
 				Usage: "create migration tables",
 				Action: func(c *cli.Context) error {
-					ctx, app, err := bunapp.StartCLI(c)
+					ctx, app, err := uptracebundle.StartCLI(c)
 					if err != nil {
 						return err
 					}
@@ -64,7 +65,7 @@ func NewCHCommand(migrations *chmigrate.Migrations) *cli.Command {
 				Name:  "migrate",
 				Usage: "migrate database",
 				Action: func(c *cli.Context) error {
-					ctx, app, err := bunapp.StartCLI(c)
+					ctx, app, err := uptracebundle.StartCLI(c)
 					if err != nil {
 						return err
 					}
@@ -90,7 +91,7 @@ func NewCHCommand(migrations *chmigrate.Migrations) *cli.Command {
 				Name:  "rollback",
 				Usage: "rollback the last migration group",
 				Action: func(c *cli.Context) error {
-					ctx, app, err := bunapp.StartCLI(c)
+					ctx, app, err := uptracebundle.StartCLI(c)
 					if err != nil {
 						return err
 					}
@@ -116,7 +117,7 @@ func NewCHCommand(migrations *chmigrate.Migrations) *cli.Command {
 				Name:  "reset",
 				Usage: "reset ClickHouse schema",
 				Action: func(c *cli.Context) error {
-					ctx, app, err := bunapp.StartCLI(c)
+					ctx, app, err := uptracebundle.StartCLI(c)
 					if err != nil {
 						return err
 					}
@@ -159,7 +160,7 @@ func NewCHCommand(migrations *chmigrate.Migrations) *cli.Command {
 				Name:  "lock",
 				Usage: "lock migrations",
 				Action: func(c *cli.Context) error {
-					ctx, app, err := bunapp.StartCLI(c)
+					ctx, app, err := uptracebundle.StartCLI(c)
 					if err != nil {
 						return err
 					}
@@ -173,7 +174,7 @@ func NewCHCommand(migrations *chmigrate.Migrations) *cli.Command {
 				Name:  "unlock",
 				Usage: "unlock migrations",
 				Action: func(c *cli.Context) error {
-					ctx, app, err := bunapp.StartCLI(c)
+					ctx, app, err := uptracebundle.StartCLI(c)
 					if err != nil {
 						return err
 					}
@@ -187,7 +188,7 @@ func NewCHCommand(migrations *chmigrate.Migrations) *cli.Command {
 				Name:  "create_go",
 				Usage: "create Go migration",
 				Action: func(c *cli.Context) error {
-					ctx, app, err := bunapp.StartCLI(c)
+					ctx, app, err := uptracebundle.StartCLI(c)
 					if err != nil {
 						return err
 					}
@@ -209,7 +210,7 @@ func NewCHCommand(migrations *chmigrate.Migrations) *cli.Command {
 				Name:  "create_sql",
 				Usage: "create up and down SQL migrations",
 				Action: func(c *cli.Context) error {
-					ctx, app, err := bunapp.StartCLI(c)
+					ctx, app, err := uptracebundle.StartCLI(c)
 					if err != nil {
 						return err
 					}
@@ -234,7 +235,7 @@ func NewCHCommand(migrations *chmigrate.Migrations) *cli.Command {
 				Name:  "status",
 				Usage: "print migrations status",
 				Action: func(c *cli.Context) error {
-					ctx, app, err := bunapp.StartCLI(c)
+					ctx, app, err := uptracebundle.StartCLI(c)
 					if err != nil {
 						return err
 					}